requirements.txt
django
south
pinax-theme-bootstrap==5.0.1
django-user-accounts==1.0b17
metron==1.3
eventlog==0.6.7
django-jsonfield==0.9.11
djangorestframework
django-oauth-toolkit
markdown
django-filter
requests
lxml
django-braces
python-openid
git+https://github.com/teddziuba/django-sslserver.git#egg=django-sslserver
